+++ packages/preset-vue-3/src/index.js
@@ -1,20 +1,29 @@
+import merge from 'lodash.merge';
 import { VueLoaderPlugin } from 'vue-loader';
 
 /**
  * Vue preset.
+ * @param   {import('./index').VuePresetOptions} [options]
  * @returns {import('./index').Preset}
  */
-export function vue() {
+export function vue(options = {}) {
+  const opts = merge(options, {
+    vue: {
+      reactivityTransform: true,
+      experimentalInlineMatchResource: true,
+    },
+    svgo: {
+      plugins: [{ prefixIds: true }, { removeViewBox: false }],
+    },
+  });
+
   return {
     name: 'vue',
     async handler(config, { extendWebpack }) {
       await extendWebpack(config, async (webpackConfig) => {
         const vueLoader = {
           loader: 'vue-loader',
-          options: {
-            reactivityTransform: true,
-            experimentalInlineMatchResource: true,
-          },
+          options: opts.vue,
         };
 
         webpackConfig.module.rules.push(
@@ -27,8 +36,16 @@ export function vue() {
             resourceQuery(input) {
               return input.includes('as-vue-component');
             },
-            use: [vueLoader, 'vue-svg-loader'],
-          }
+            use: [
+              vueLoader,
+              {
+                loader: 'vue-svg-loader',
+                options: {
+                  svgo: opts.svgo,
+                },
+              },
+            ],
+          },
         );
 
         webpackConfig.plugins.push(new VueLoaderPlugin());